Move redirection test up
